Introduce HTTP easy interface.

+++ src/net/tcp.c
@@ -22,7 +22,7 @@ extern const t_rinoosocket_class socket_class_tcp;
  *
  * @return Socket pointer on success or NULL if an error occurs
  */
-t_rinoosocket *rinoo_tcp_client(t_rinoosched *sched, t_ip ip, uint32_t port, uint32_t timeout)
+t_rinoosocket *rinoo_tcp_client(t_rinoosched *sched, t_ip ip, uint16_t port, uint32_t timeout)
 {
 	t_rinoosocket *socket;
 	struct sockaddr_in addr;
@@ -54,7 +54,7 @@ t_rinoosocket *rinoo_tcp_client(t_rinoosched *sched, t_ip ip, uint32_t port, uin
  *
  * @return Socket pointer to the server on success or NULL if an error occurs
  */
-t_rinoosocket *rinoo_tcp_server(t_rinoosched *sched, t_ip ip, uint32_t port)
+t_rinoosocket *rinoo_tcp_server(t_rinoosched *sched, t_ip ip, uint16_t port)
 {
 	t_rinoosocket *socket;
 	struct sockaddr_in addr;
@@ -82,7 +82,7 @@ t_rinoosocket *rinoo_tcp_server(t_rinoosched *sched, t_ip ip, uint32_t port)
  *
  * @return A pointer to the new socket on success or NULL if an error occurs
  */
-t_rinoosocket *rinoo_tcp_accept(t_rinoosocket *socket, t_ip *fromip, uint32_t *fromport)
+t_rinoosocket *rinoo_tcp_accept(t_rinoosocket *socket, t_ip *fromip, uint16_t *fromport)
 {
 	t_rinoosocket *new;
 	struct sockaddr_in addr;
